requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.23
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pypdf>=3.17.0
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1